Move DB query functions to appropriate classes
