+++ openlibrary/admin/vitals.py
@@ -2,6 +2,8 @@ import yaml
 from statsd import StatsClient
 
 from openlibrary.core import db
+from openlibrary.core.bookshelves import Bookshelves
+from openlibrary.core.ratings import Ratings
 
 STATSD_EVENT_PREFIX = "stats.ol.participation"
 
@@ -73,70 +75,9 @@ def write_to_statsd(configfile, stats):
 
 def gather_participation_scores() -> dict[str, int | dict[str, int]]:
     results = {}
-    results.update(calc_star_rating_counts())
-    results.update(calc_reading_log_counts())
+    results.update(Ratings.calc_star_rating_counts())
+    results.update(Bookshelves.calc_reading_log_counts())
     results.update(calc_list_counts())
-    return results
-
-def calc_star_rating_counts() -> dict[str, int]:
-    results = {}
-    oldb = db.get_db()
-    total_ratings_query = """
-        select count(*) from ratings
-        WHERE created >= date_trunc('hour', now() - interval '1 hour')
-            AND created <  date_trunc('hour', now());
-    """
-    totals = oldb.query(total_ratings_query)
-    results['star_ratings'] = next(iter(totals))['count']
-
-    distinct_raters_query = """
-        select count(distinct username) from ratings
-        WHERE created >= date_trunc('hour', now() - interval '1 hour')
-            AND created <  date_trunc('hour', now());
-    """
-    distinct_totals = oldb.query(distinct_raters_query)
-    results["distinct_star_ratings"] = next(iter(distinct_totals))['count']
-
-    return results
-
-def calc_reading_log_counts() -> dict[str, dict[str, int]]:
-    def normalize_shelf_name(shelf_name: str) -> str:
-        return shelf_name.lower().replace(" ", "_")
-
-    results = {}
-    oldb = db.get_db()
-    total_books_logged_query = """
-        select count(bb.bookshelf_id) as cnt,
-               b.name as bookshelf_name
-        from bookshelves b
-        left join bookshelves_books bb
-            on bb.bookshelf_id = b.id
-            and bb.created >= date_trunc('hour', now() - interval '1 hour')
-            and bb.created <  date_trunc('hour', now())
-        group by b.id, b.name
-        order by b.id;
-    """
-    totals = oldb.query(total_books_logged_query)
-    results['reading_logs'] = {
-        normalize_shelf_name(i.bookshelf_name): i.cnt for i in totals
-    }
-
-    distinct_readers_logging_query = """
-        select count(distinct bb.username) as cnt,
-           b.name as bookshelf_name
-        from bookshelves b
-        left join bookshelves_books bb
-            on bb.bookshelf_id = b.id
-            and bb.created >= date_trunc('hour', now() - interval '1 hour')
-            and bb.created <  date_trunc('hour', now())
-        group by b.id, b.name
-        order by b.id;
-    """
-    distinct_totals = oldb.query(distinct_readers_logging_query)
-    results["distinct_reading_logs"] = {
-        normalize_shelf_name(i.bookshelf_name): i.cnt for i in distinct_totals
-    }
-
     return results
 
 def calc_list_counts() -> dict[str, int]:
@@ -161,6 +102,3 @@ def calc_list_counts() -> dict[str, int]:
     results["distinct_list_counts"] = next(iter(distinct_totals))['distinct_list_creators']
 
     return results
-
-def calc_edit_counts() -> dict[str, str]:
-    raise NotImplementedError()
